# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MedConsent Pro API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /specialties:
    get:
      operationId: listSpecialties
      tags: [specialties]
      summary: List specialties
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: activeOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of specialties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Specialty"
    post:
      operationId: createSpecialty
      tags: [specialties]
      summary: Create a specialty
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialtyInput"
      responses:
        "201":
          description: Specialty created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialty"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /specialties/{id}:
    get:
      operationId: getSpecialty
      tags: [specialties]
      summary: Get specialty by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Specialty found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialty"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateSpecialty
      tags: [specialties]
      summary: Update specialty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SpecialtyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Specialty"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteSpecialty
      tags: [specialties]
      summary: Delete specialty
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: ["string", "null"]
...
    Specialty:
      type: object
      required: [id, nameAr, isActive, createdAt, updatedAt]
      properties:
        id:
          type: integer
        nameAr:
          type: string
        nameEn:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
    SpecialtyInput:
      type: object
      required: [nameAr]
      properties:
        nameAr:
          type: string
          minLength: 1
        nameEn:
          type: string
        isActive:
          type: boolean
    SpecialtyUpdate:
      type: object
      properties:
        nameAr:
          type: string
        nameEn:
          type: string
        isActive:
          type: boolean
...