# lib/api-zod/src/generated/types/specialtyUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MedConsent Pro API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SpecialtyUpdate {
  nameAr?: string;
  nameEn?: string;
  isActive?: boolean;
}